Move preview caching test to `no_concurrency` suite (#645)
