+++ integration_tests/no_concurrency/sdk/utils.py
@@ -1,0 +1,5 @@
+SENTIMENT_SQL_QUERY = "select * from hotel_reviews"
+
+
+def get_integration_name() -> str:
+    return "aqueduct_demo"
